Create Power Of Two Integers.cpp
Power Of Two Integers
Asked in:  
Housing
Amazon
Given a positive integer which fits in a 32 bit signed integer, find if it can be expressed as A^P where P > 1 and A > 0. A and P both should be integers.

Example

Input : 4
Output : True  
as 2^2 = 4. 
Output Status:
Correct Answer. You got 250/250 points!